refactor: gas optimization — struct packing and library externalization

+++ loan-provider/script/deployment/local/SchedulePhase3Local.s.sol
@@ -16,7 +16,7 @@ import {DeploymentConstants} from "../DeploymentConstants.sol";
 /// @custom:security Only for local Anvil deployments (chainId 31337)
 contract SchedulePhase3Local is Script, DeploymentHelper {
     // ===== Constants =====
-    uint256 constant STRATEGY_CAP = type(uint256).max;
+    uint256 constant STRATEGY_CAP = type(uint96).max;
 
     // ===== Config =====
     HelperConfig public helperConfig;
